Migrate matrix-vector multiplication tests
