+++ src/webgpu/shader/execution/expression/expression.ts
@@ -1,7 +1,7 @@
 import { globalTestConfig } from '../../../../common/framework/test_config.js';
 import { objectEquals, unreachable } from '../../../../common/util/util.js';
 import { GPUTest } from '../../../gpu_test.js';
-import { compare, Comparator, anyOf } from '../../../util/compare.js';
+import { compare, Comparator } from '../../../util/compare.js';
 import {
   ScalarType,
   Scalar,
@@ -11,20 +11,16 @@ import {
   Value,
   Vector,
   VectorType,
-  f32,
   u32,
   i32,
   Matrix,
   MatrixType,
   ScalarBuilder,
 } from '../../../util/conversion.js';
-import { MatrixVectorToVector, VectorMatrixToVector } from '../../../util/f32_interval.js';
 import { FPInterval } from '../../../util/floating_point.js';
 import {
   cartesianProduct,
-  map2DArray,
   QuantizeFunc,
-  quantizeToF32,
   quantizeToI32,
   quantizeToU32,
 } from '../../../util/math.js';
@@ -802,118 +798,6 @@ function packScalarsToVector(
 export type IntervalFilter =
   | 'finite' // Expected to be finite in the interval numeric space
   | 'unfiltered'; // No expectations
-
-/**
- * @returns a Case for the params and the vector of intervals generator provided
- * @param mat the matrix param to pass in
- * @param vec the vector to pass in
- * @param filter what interval filtering to apply
- * @param ops callbacks that implement generating a vector of acceptance
- *            intervals for a matrix and a vector.
- */
-function makeMatrixVectorToVectorCase(
-  mat: number[][],
-  vec: number[],
-  filter: IntervalFilter,
-  ...ops: MatrixVectorToVector[]
-): Case | undefined {
-  mat = map2DArray(mat, quantizeToF32);
-  vec = vec.map(quantizeToF32);
-  const mat_f32 = map2DArray(mat, f32);
-  const vec_f32 = vec.map(f32);
-
-  const results = ops.map(o => o(mat, vec));
-  if (filter === 'finite' && results.some(v => v.some(e => !e.isFinite()))) {
-    return undefined;
-  }
-  return {
-    input: [new Matrix(mat_f32), new Vector(vec_f32)],
-    expected: anyOf(...results),
-  };
-}
-
-/**
- * @returns an array of Cases for operations over a range of inputs
- * @param mats array of inputs to try for the matrix input
- * @param vecs array of inputs to try for the vector input
- * @param filter what interval filtering to apply
- * @param ops callbacks that implement generating a vector of acceptance
- *            intervals for a matrix and a vector.
- */
-export function generateMatrixVectorToVectorCases(
-  mats: number[][][],
-  vecs: number[][],
-  filter: IntervalFilter,
-  ...ops: MatrixVectorToVector[]
-): Case[] {
-  // Cannot use cartesianProduct here, due to heterogeneous types
-  const cases: Case[] = [];
-  mats.forEach(mat => {
-    vecs.forEach(vec => {
-      const c = makeMatrixVectorToVectorCase(mat, vec, filter, ...ops);
-      if (c !== undefined) {
-        cases.push(c);
-      }
-    });
-  });
-  return cases;
-}
-
-/**
- * @returns a Case for the params and the vector of intervals generator provided
- * @param vec the vector to pass in
- * @param mat the matrix param to pass in
- * @param filter what interval filtering to apply
- * @param ops callbacks that implement generating a vector of acceptance
- *            intervals for a vector and a matrix.
- */
-function makeVectorMatrixToVectorCase(
-  vec: number[],
-  mat: number[][],
-  filter: IntervalFilter,
-  ...ops: VectorMatrixToVector[]
-): Case | undefined {
-  vec = vec.map(quantizeToF32);
-  mat = map2DArray(mat, quantizeToF32);
-  const vec_f32 = vec.map(f32);
-  const mat_f32 = map2DArray(mat, f32);
-
-  const results = ops.map(o => o(vec, mat));
-  if (filter === 'finite' && results.some(v => v.some(e => !e.isFinite()))) {
-    return undefined;
-  }
-  return {
-    input: [new Vector(vec_f32), new Matrix(mat_f32)],
-    expected: anyOf(...results),
-  };
-}
-
-/**
- * @returns an array of Cases for operations over a range of inputs
- * @param vecs array of inputs to try for the vector input
- * @param mats array of inputs to try for the matrix input
- * @param filter what interval filtering to apply
- * @param ops callbacks that implement generating a vector of acceptance
- *            intervals for a vector and a matrix.
- */
-export function generateVectorMatrixToVectorCases(
-  vecs: number[][],
-  mats: number[][][],
-  filter: IntervalFilter,
-  ...ops: VectorMatrixToVector[]
-): Case[] {
-  // Cannot use cartesianProduct here, due to heterogeneous types
-  const cases: Case[] = [];
-  vecs.forEach(vec => {
-    mats.forEach(mat => {
-      const c = makeVectorMatrixToVectorCase(vec, mat, filter, ...ops);
-      if (c !== undefined) {
-        cases.push(c);
-      }
-    });
-  });
-  return cases;
-}
 
 /**
  * A function that performs a binary operation on x and y, and returns the expected
